requirements.txt
# ============================================================================
# FloraWatch - Requirements.txt
# Dependencias completas para el proyecto de monitoreo de floración global
# ============================================================================

# -------------------------
# FRAMEWORK WEB PRINCIPAL
# -------------------------
Flask==3.0.0                    # Framework web principal
Flask-CORS==4.0.0              # Soporte para CORS (Cross-Origin Resource Sharing)
Werkzeug==3.0.1                # Utilidades WSGI para Flask
Jinja2==3.1.2                  # Motor de plantillas

# -------------------------
# ANÁLISIS DE DATOS CORE
# -------------------------
numpy==1.25.2                  # Computación numérica fundamental
pandas==2.1.1                  # Manipulación y análisis de datos
scipy==1.11.3                  # Algoritmos científicos y estadísticos
scikit-learn==1.3.1            # Machine Learning y análisis predictivo

# -------------------------
# VISUALIZACIÓN Y GRÁFICOS
# -------------------------
matplotlib==3.8.0              # Gráficos estáticos
plotly==5.17.0                 # Gráficos interactivos
seaborn==0.12.2                # Visualización estadística
folium==0.14.0                 # Mapas interactivos

# -------------------------
# PROCESAMIENTO GEOESPACIAL
# -------------------------
rasterio==1.3.9                # Lectura/escritura de datos raster
geopandas==0.14.0              # Análisis de datos geoespaciales
shapely==2.0.2                 # Manipulación de geometrías
pyproj==3.6.1                  # Transformaciones cartográficas
Fiona==1.9.5                   # I/O de archivos vectoriales geoespaciales

# -------------------------
# COMUNICACIÓN HTTP/APIs
# -------------------------
requests==2.31.0               # Cliente HTTP
urllib3==2.0.6                 # Biblioteca HTTP de bajo nivel
certifi==2023.7.22            # Certificados SSL/TLS

# -------------------------
# PROCESAMIENTO DE FECHAS
# -------------------------
python-dateutil==2.8.2        # Extensiones para datetime
pytz==2023.3                  # Zonas horarias

# -------------------------
# PROCESAMIENTO DE IMÁGENES
# -------------------------
Pillow==10.0.1                 # Biblioteca de imágenes Python (PIL fork)
imageio==2.31.5               # Lectura/escritura de imágenes

# -------------------------
# CONFIGURACIÓN Y ENTORNO
# -------------------------
python-dotenv==1.0.0          # Carga variables de entorno desde .env
click==8.1.7                  # Interfaz de línea de comandos
colorama==0.4.6               # Colores en terminal (Windows)

# -------------------------
# SERIALIZACIÓN Y FORMATOS
# -------------------------
lxml==4.9.3                   # Procesamiento XML/HTML
beautifulsoup4==4.12.2        # Parsing HTML/XML
openpyxl==3.1.2               # Archivos Excel

# -------------------------
# UTILIDADES DE DESARROLLO
# -------------------------
tqdm==4.66.1                  # Barras de progreso
joblib==1.3.2                 # Paralelización y persistencia

# -------------------------
# BASE DE DATOS (OPCIONAL)
# -------------------------
SQLAlchemy==2.0.21            # ORM de base de datos
Flask-SQLAlchemy==3.1.1       # Integración SQLAlchemy con Flask
psycopg2-binary==2.9.7       # Driver PostgreSQL

# -------------------------
# CACHE Y REDIS (OPCIONAL)
# -------------------------
redis==5.0.1                  # Cliente Redis para cache

# -------------------------
# AUTENTICACIÓN (OPCIONAL)
# -------------------------
Flask-Login==0.6.3            # Gestión de sesiones de usuario
PyJWT==2.8.0                  # JSON Web Tokens

# -------------------------
# SERVIDOR DE PRODUCCIÓN
# -------------------------
gunicorn==21.2.0              # Servidor WSGI para producción
gevent==23.7.0                # Concurrencia asíncrona

# -------------------------
# TESTING Y CALIDAD DE CÓDIGO
# -------------------------
pytest==7.4.2                 # Framework de testing
pytest-cov==4.1.0            # Cobertura de código
pytest-flask==1.3.0          # Testing específico para Flask
black==23.9.1                 # Formateador de código
flake8==6.1.0                 # Linter de código
isort==5.12.0                 # Ordenador de imports

# -------------------------
# DOCUMENTACIÓN
# -------------------------
Sphinx==7.2.6                # Generador de documentación

# -------------------------
# MONITOREO Y LOGGING
# -------------------------
structlog==23.1.0            # Logging estructurado

# -------------------------
# SEGURIDAD
# -------------------------
cryptography==41.0.7         # Primitivas criptográficas

# -------------------------
# DEPENDENCIAS DEL SISTEMA
# -------------------------
# Nota: GDAL se instala mejor desde conda o usando wheels precompilados
# GDAL==3.7.3                # Abstraction Library for Geospatial Data
# Para instalación más fácil, usar: conda install -c conda-forge gdal